PP-14489 - Merged test and added IT test for currentStatus is REFUND_ERROR
